pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evolution-suite"
version = "0.1.0"
description = "RTS-style command center for autonomous AI agent pools that evolve codebases"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Reagan Triminio" }
]
keywords = ["ai", "agents", "automation", "claude", "evolution"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "watchfiles>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[project.scripts]
evolution = "evolution_suite.cli:app"

[project.urls]
Homepage = "https://github.com/reagantriminio/evolution-suite"
Repository = "https://github.com/reagantriminio/evolution-suite"

[tool.hatch.build.targets.wheel]
packages = ["evolution_suite"]

[tool.hatch.build.targets.wheel.shared-data]
"templates" = "share/evolution-suite/templates"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
